backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
opensearch-py
psycopg2-binary
pandas
pandera
python-dotenv
sentence-transformers
numpy
SQLAlchemy
